refactor: remove unused assets
