+++ ts/features/itwallet/common/components/ItwCredentialCard/ItwDigitalVersionBadge.tsx
@@ -5,16 +5,15 @@ import {
   IOColors,
   makeFontStyleObject
 } from "@pagopa/io-app-design-system";
-import { memo } from "react";
 import Color from "color";
+import { memo } from "react";
 import { Platform, StyleSheet, Text, View } from "react-native";
-import I18n from "../../../../i18n";
-
-export type TagColorScheme = "default" | "faded" | "greyscale";
+import I18n from "../../../../../i18n";
+import { CardColorScheme } from "./types";
 
 type DigitalVersionBadgeProps = {
   credentialType: string;
-  colorScheme: TagColorScheme;
+  colorScheme: CardColorScheme;
 };
 
 type CredentialTypesProps = {
@@ -24,7 +23,7 @@ type CredentialTypesProps = {
 
 const getColorPropsByScheme = (
   credentialType: string,
-  colorScheme: TagColorScheme
+  colorScheme: CardColorScheme
 ) => {
   const mapCredentialTypes: Record<string, CredentialTypesProps> = {
     MDL: {
