+++ ts/features/itwallet/common/components/ItwCredentialCard/types.ts
@@ -1,0 +1,1 @@
+export type CardColorScheme = "default" | "faded" | "greyscale";
